(PC-32696)[PRO] refactor: relocate test tile, re-order commons
