+++ pro/src/components/IndividualOffer/UsefulInformationScreen/utils.ts
@@ -1,11 +1,9 @@
 import {
   GetIndividualOfferResponseModel,
   type GetIndividualOfferWithAddressResponseModel,
-  SubcategoryResponseModel,
   VenueListItemResponseModel,
 } from 'apiClient/v1'
 import {
-  CATEGORY_STATUS,
   OFFER_STATUS_PENDING,
   OFFER_STATUS_REJECTED,
 } from 'commons/core/Offers/constants'
@@ -17,25 +15,6 @@ import { computeAddressDisplayName } from 'repository/venuesService'
 
 import { DEFAULT_USEFULL_INFORMATION_INTITIAL_VALUES } from './constants'
 import { UsefulInformationFormValues } from './types'
-
-export const getFilteredVenueListBySubcategory = (
-  venueList: VenueListItemResponseModel[],
-  subcategory?: SubcategoryResponseModel
-): VenueListItemResponseModel[] => {
-  if (!subcategory) {
-    return venueList
-  }
-
-  if (subcategory.onlineOfflinePlatform === CATEGORY_STATUS.ONLINE_OR_OFFLINE) {
-    return venueList
-  }
-
-  return venueList.filter((venue) =>
-    subcategory.onlineOfflinePlatform === CATEGORY_STATUS.ONLINE
-      ? venue.isVirtual
-      : !venue.isVirtual
-  )
-}
 
 interface SetDefaultInitialValuesFromOfferProps {
   offer: GetIndividualOfferWithAddressResponseModel
