+++ pro/src/pages/IndividualOffer/IndividualOfferDetailsAndInformations/IndividualOfferDetails/components/DetailsEanSearch/DetailsEanSearch.tsx
@@ -8,7 +8,7 @@ import { selectCurrentOffererId } from 'commons/store/user/selectors'
 import { FormLayout } from 'components/FormLayout/FormLayout'
 import strokeBarcode from 'icons/stroke-barcode.svg'
 import { Product } from 'pages/IndividualOffer/IndividualOfferDetailsAndInformations/IndividualOfferDetails/commons/types'
-import { isSubCategoryCD } from 'pages/IndividualOffer/IndividualOfferDetailsAndInformations/IndividualOfferDetails/commons/utils/utils'
+import { isSubCategoryCD } from 'pages/IndividualOffer/IndividualOfferDetailsAndInformations/IndividualOfferDetails/commons/utils'
 import { eanSearchValidationSchema } from 'pages/IndividualOffer/IndividualOfferDetailsAndInformations/IndividualOfferDetails/commons/validationSchema'
 import { EanSearchCallout } from 'pages/IndividualOffer/IndividualOfferDetailsAndInformations/IndividualOfferDetails/components/EanSearchCallout/EanSearchCallout'
 import { Button } from 'ui-kit/Button/Button'
