+++ pro/src/pages/IndividualOffer/IndividualOfferDetailsAndInformations/IndividualOfferDetails/components/DetailsForm/SuggestedSubcategories/SuggestedSubcategories.tsx
@@ -11,7 +11,7 @@ import {
   onCategoryChange,
   buildCategoryOptions,
   buildSubcategoryOptions,
-} from 'pages/IndividualOffer/IndividualOfferDetailsAndInformations/IndividualOfferDetails/commons/utils/utils'
+} from 'pages/IndividualOffer/IndividualOfferDetailsAndInformations/IndividualOfferDetails/commons/utils'
 import { RadioButton } from 'ui-kit/form/RadioButton/RadioButton'
 import { Select } from 'ui-kit/form/Select/Select'
 import { InfoBox } from 'ui-kit/InfoBox/InfoBox'
