+++ pro/src/pages/IndividualOffer/IndividualOfferDetailsAndInformations/IndividualOfferInformationsLegacy/InformationsScreen/IndividualOfferForm/Categories/Categories.tsx
@@ -12,17 +12,17 @@ import { Callout } from 'components/Callout/Callout'
 import { CalloutVariant } from 'components/Callout/types'
 import { FormLayout } from 'components/FormLayout/FormLayout'
 import fullMoreIcon from 'icons/full-more.svg'
+import { getFilteredVenueListBySubcategory } from 'pages/IndividualOffer/commons/getFilteredVenueList'
 import { IndividualOfferFormValues } from 'pages/IndividualOffer/commons/types'
 import {
   SUBCATEGORIES_FIELDS_DEFAULT_VALUES,
   FORM_DEFAULT_VALUES,
 } from 'pages/IndividualOffer/IndividualOfferDetailsAndInformations/commons/constants'
-import { getFilteredVenueListBySubcategory } from 'pages/IndividualOffer/IndividualOfferDetailsAndInformations/IndividualOfferDetails/commons/getFilteredVenueList'
 import {
   hasMusicType,
   buildCategoryOptions,
   buildSubcategoryOptions,
-} from 'pages/IndividualOffer/IndividualOfferDetailsAndInformations/IndividualOfferDetails/commons/utils/utils'
+} from 'pages/IndividualOffer/IndividualOfferDetailsAndInformations/IndividualOfferDetails/commons/utils'
 import { Select } from 'ui-kit/form/Select/Select'
 import { InfoBox } from 'ui-kit/InfoBox/InfoBox'
