+++ pro/src/pages/IndividualOffer/commons/filterCategories.ts
