+++ pro/src/pages/IndividualOffer/commons/getFilteredVenueList.ts
